Undone the CSSNumericValue patch. What a stupid idea.

+++ liblayout/source/org/jfree/layouting/input/style/values/CSSTimeValue.java
@@ -36,7 +36,7 @@ package org.jfree.layouting.input.style.values;
  */
 public class CSSTimeValue extends CSSNumericValue
 {
-  public CSSTimeValue(final CSSTimeType type, final long value)
+  public CSSTimeValue(final CSSTimeType type, final double value)
   {
     super(type, value);
   }
